Move factory function so that it can use -update flag
