+++ enterprise/internal/campaigns/resolvers/main_test.go
@@ -2,11 +2,17 @@ package resolvers
 
 import (
 	"flag"
+	"net/http"
 	"os"
+	"path/filepath"
+	"strings"
 	"testing"
 
+	"github.com/dnaeon/go-vcr/cassette"
 	"github.com/inconshreveable/log15"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
+	"github.com/sourcegraph/sourcegraph/internal/httpcli"
+	"github.com/sourcegraph/sourcegraph/internal/httptestutil"
 )
 
 func init() {
@@ -21,4 +27,37 @@ func TestMain(m *testing.M) {
 		log15.Root().SetHandler(log15.DiscardHandler())
 	}
 	os.Exit(m.Run())
+}
+
+func newGithubClientFactory(t testing.TB, name string) (*httpcli.Factory, func()) {
+	t.Helper()
+
+	cassete := filepath.Join("testdata/vcr/", strings.Replace(name, " ", "-", -1))
+
+	rec, err := httptestutil.NewRecorder(cassete, *update, func(i *cassette.Interaction) error {
+		return nil
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	mw := httpcli.NewMiddleware(githubProxyRedirectMiddleware)
+
+	hc := httpcli.NewFactory(mw, httptestutil.NewRecorderOpt(rec))
+
+	return hc, func() {
+		if err := rec.Stop(); err != nil {
+			t.Errorf("failed to update test data: %s", err)
+		}
+	}
+}
+
+func githubProxyRedirectMiddleware(cli httpcli.Doer) httpcli.Doer {
+	return httpcli.DoerFunc(func(req *http.Request) (*http.Response, error) {
+		if req.URL.Hostname() == "github-proxy" {
+			req.URL.Host = "api.github.com"
+			req.URL.Scheme = "https"
+		}
+		return cli.Do(req)
+	})
 }
